feat: Add ngrok-skip-browser-warning header to CORS setup
This commit adds the "ngrok-skip-browser-warning" header to the CORS setup in the backend app. This header is set to "true" to skip the browser warning when using ngrok for local development. This improves the developer experience by removing the need to manually dismiss the warning every time the app is accessed through ngrok.